requirements.txt
python-dotenv
openai
bs4
requests
rich
ollama